+++ src/bootstrap/basstub.c
@@ -121,7 +121,7 @@ int main(int argc, char **argv) {
     if (vm_has_error(boot->vm)) {
         BppError run_err = vm_get_error(boot->vm);
         VDevContext *vdev = vm_get_vdev(boot->vm);
-        vdev_printf(vdev, "?Error %d: %s in line %g\n", run_err.code, run_err.message, run_err.line);
+        vdev_printf(vdev, "?Error %d: %s in line %lld\n", run_err.code, run_err.message, (long long)run_err.line);
         boot_shutdown(boot);
         return 1;
     }
